Open landmask file with parallel access
This is done because the darshan I/O profiler will misinterpret the I/O
to be POSIX if the first NetCDF file opened in the program is not opened
with parallel access.

Although, it is probably a good idea to open the landmask file in
parallel as this will hurt performance when running with a high number
of ranks.